apps/api/requirements.txt
# Study2026 API Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0  # PostgreSQL 异步驱动
psycopg2-binary==2.9.9  # PostgreSQL 同步驱动 (可选)

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI/ML
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langchain-community==0.0.10
openai==1.7.2
anthropic==0.8.1
tiktoken==0.5.2
numpy==1.26.3

# Vector Search (可选 - 使用更高级的向量数据库时)
chromadb==0.4.22
qdrant-client==1.7.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1

# Email
jinja2==3.1.3

# Task Scheduling
apscheduler==3.10.4

# Redis (缓存和消息队列)
redis==5.0.1

# HTTP Client
aiohttp==3.9.1
requests==2.31.0
httpx==0.26.0

# Image Processing
opencv-python==4.8.1.78
pillow==10.1.0
scikit-image==0.22.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenClaw Integration
pydantic==2.5.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
